Adds image and co-author to author show page
